requirements.txt
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.1
google-api-python-client==2.146.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1